Withdraw ix

+++ programs/single-asset-lending-program/src/state.rs
@@ -10,7 +10,7 @@ pub struct VaultState {
   pub total_deposits: u64,
   pub interest_rate: f32,
   pub borrow_percentage_per_user: f32,
-  pub reward_factor: u64,
+  pub reward_factor: f64,
 }
 
 impl VaultState {
